Exit as soon as possible if database writes return ENOSPC (Debian bug #167159). * src/check_mandirs.c (testmandirs, create_db, update_db,   purge_normal, purge_whatis, check_multi_key, purge_missing):   Return int rather than short, to make space for -errno. * src/mandb.c (update_one_file, update_db_wrapper, mandb,   process_manpath): Likewise. * src/check_mandirs.h (create_db, update_db, purge_missing): Update   prototypes. * src/check_mandirs.c (create_db): Return -errno for database open   errors other than EACCES and EROFS. * src/mandb.c (mandb): Return early on errors from create_db and   update_db_wrapper. Propagate errors from xcopy as well as   returning early.   (process_manpath): Return early on errors from mandb.   (main): Exit FATAL on errors from process_manpath.

+++ src/check_mandirs.h
@@ -26,7 +26,7 @@
 extern void test_manfile (const char *file, const char *path);
 extern void update_db_time (void);
 extern void reset_db_time (void);
-extern short create_db (const char *manpath);
-extern short update_db (const char *manpath);
+extern int create_db (const char *manpath);
+extern int update_db (const char *manpath);
 extern void purge_pointers (const char *name);
-extern short purge_missing (const char *manpath, const char *catpath);
+extern int purge_missing (const char *manpath, const char *catpath);
